cmds/core/netcat: separate plan9 impl. more granular
