+++ cmds/core/netcat/connect.go
@@ -7,16 +7,19 @@ package main
 import (
 	"bufio"
 	"bytes"
+	"crypto/tls"
 	"errors"
 	"fmt"
 	"io"
 	"log"
+	"net"
 	"net/url"
 	"sync"
 	"time"
 
 	"golang.org/x/net/proxy"
 
+	"github.com/mdlayher/vsock"
 	"github.com/u-root/u-root/pkg/netcat"
 )
 
@@ -80,6 +83,102 @@ func (c *cmd) connectMode(output io.Writer, network, address string) error {
 	wg.Wait()
 
 	return nil
+}
+
+func (c *cmd) establishConnection(network, address string) (net.Conn, error) {
+	var (
+		err  error
+		conn net.Conn
+	)
+
+	dialer := &net.Dialer{
+		Timeout: c.config.Timing.Wait,
+	}
+
+	if c.config.ConnectionModeOptions.SourceHost != "" {
+		switch c.config.ProtocolOptions.SocketType {
+
+		case netcat.SOCKET_TYPE_TCP:
+			dialer.LocalAddr, err = net.ResolveTCPAddr(network, fmt.Sprintf("%v:%v", c.config.ConnectionModeOptions.SourceHost, c.config.ConnectionModeOptions.SourcePort))
+			if err != nil {
+				return nil, fmt.Errorf("failed to resolve source address %v", err)
+			}
+
+		case netcat.SOCKET_TYPE_UDP:
+			dialer.LocalAddr, err = net.ResolveUDPAddr(network, fmt.Sprintf("%v:%v", c.config.ConnectionModeOptions.SourceHost, c.config.ConnectionModeOptions.SourcePort))
+			if err != nil {
+				return nil, fmt.Errorf("failed to resolve source address %v", err)
+			}
+
+		case netcat.SOCKET_TYPE_UNIX:
+			dialer.LocalAddr, err = net.ResolveUnixAddr(network, c.config.ConnectionModeOptions.SourceHost)
+			if err != nil {
+				return nil, fmt.Errorf("failed to resolve source address %v", err)
+			}
+
+		case netcat.SOCKET_TYPE_UDP_UNIX:
+			dialer.LocalAddr, err = net.ResolveUnixAddr(network, c.config.ConnectionModeOptions.SourceHost)
+			if err != nil {
+				return nil, fmt.Errorf("failed to resolve source address %v", err)
+			}
+
+		case netcat.SOCKET_TYPE_SCTP:
+			return connectToSCTPSocket(network, address)
+
+		case netcat.SOCKET_TYPE_VSOCK:
+			cid, port, err := netcat.SplitVSockAddr(address)
+			if err != nil {
+				return nil, fmt.Errorf("failed to resolve VSOCK address: %v", err)
+			}
+
+			return vsock.Dial(cid, port, nil)
+
+		// unsupported socket types
+		case netcat.SOCKET_TYPE_UDP_VSOCK:
+			return nil, fmt.Errorf("currently unsupported socket type %q", c.config.ProtocolOptions.SocketType)
+
+		case netcat.SOCKET_TYPE_NONE:
+		default:
+			return nil, fmt.Errorf("undefined socket type %q", c.config.ProtocolOptions.SocketType)
+		}
+	}
+
+	// Proxy Support
+	if c.config.ProxyConfig.Enabled {
+		proxyDialer, err := c.proxyDialer(dialer)
+		if err != nil {
+			return nil, err
+		}
+
+		conn, err = proxyDialer.Dial(network, address)
+		if err != nil {
+			return nil, err
+		}
+	} else {
+		// TLS Support
+		if c.config.SSLConfig.Enabled || c.config.SSLConfig.VerifyTrust {
+			tlsConfig, err := c.config.SSLConfig.GenerateTLSConfiguration()
+			if err != nil {
+				return nil, err
+			}
+
+			conn, err = tls.DialWithDialer(dialer, network, address, tlsConfig)
+			if err != nil {
+				return nil, err
+			}
+		} else {
+			conn, err = dialer.Dial(network, address)
+			if err != nil {
+				return nil, err
+			}
+		}
+	}
+
+	if c.config.Timing.Timeout > 0 {
+		conn.SetDeadline(time.Now().Add(c.config.Timing.Timeout))
+	}
+
+	return conn, nil
 }
 
 func (c *cmd) scanPorts() error {
