+++ cmds/core/netcat/connect_others.go
@@ -2,13 +2,22 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
+//go:build linux || windows
+
 package main
 
 import (
 	"fmt"
 	"net"
+
+	"github.com/ishidawataru/sctp"
 )
 
 func connectToSCTPSocket(network, address string) (net.Conn, error) {
-	return nil, fmt.Errorf("sctp is not supported on Plan 9")
+	sctpAddr, err := sctp.ResolveSCTPAddr(network, address)
+	if err != nil {
+		return nil, fmt.Errorf("failed to resolve SCTP address: %w", err)
+	}
+
+	return sctp.DialSCTP(network, nil, sctpAddr)
 }
